Implement Metrics entity and MetricsRepository
