+++ src/main/java/com/example/echo_api/persistence/dto/response/profile/ProfileDTO.java
@@ -28,10 +28,6 @@ public record ProfileDTO(
     String location,
     @JsonProperty("avatar_url") String avatarUrl,
     @JsonProperty("banner_url") String bannerUrl,
-    @JsonProperty("following_count") int followingCount,
-    @JsonProperty("follower_count") int followerCount,
-    @JsonProperty("post_count") int postCount,
-    @JsonProperty("media_count") int mediaCount,
     @JsonProperty("created_at") String createdAt
 ) {
 
@@ -42,10 +38,6 @@ public record ProfileDTO(
         String location,
         String avatarUrl,
         String bannerUrl,
-        int followingCount,
-        int followerCount,
-        int postCount,
-        int mediaCount,
         LocalDateTime createdAt
     ) {
         this(
@@ -55,10 +47,6 @@ public record ProfileDTO(
             location,
             avatarUrl,
             bannerUrl,
-            followingCount,
-            followerCount,
-            postCount,
-            mediaCount,
             createdAt != null ? createdAt.toString() : null
         );
     }
