+++ src/main/java/com/example/echo_api/persistence/mapper/ProfileMapper.java
@@ -19,10 +19,6 @@ public class ProfileMapper {
             profile.getLocation(),
             profile.getAvatarUrl(),
             profile.getBannerUrl(),
-            profile.getFollowingCount(),
-            profile.getFollowingCount(),
-            profile.getPostCount(),
-            profile.getMediaCount(),
             profile.getCreatedAt());
     }
