+++ src/main/java/com/example/echo_api/persistence/repository/MetricsRepository.java
@@ -1,0 +1,11 @@
+package com.example.echo_api.persistence.repository;
+
+import java.util.UUID;
+
+import org.springframework.data.repository.ListCrudRepository;
+
+import com.example.echo_api.persistence.model.profile.Metrics;
+
+public interface MetricsRepository extends ListCrudRepository<Metrics, UUID> {
+
+}
